Implemented persistent saving and loading
Massive work here, I've got an ugly (atm) impl. of making it save the
index object(s) and documents to separate files on disk.

If you then create a new PersistantDirectory type from the
DirectoryFactory (as done in the tests) it successfully loads them, and
it's possible to then search the persisting data

It's quite hacky at the moment though, lots of room for improvements.
But happy it works